test(gotrue): move amr claim regression test into gotrue_mfa_api_test.dart
Addresses review feedback to keep the mfa test suite in a single file.
